% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ethnic_vote.R
\name{extract_ethnic_vote}
\alias{extract_ethnic_vote}
\title{Extracting quantities of interest from computer simulations}
\usage{
extract_ethnic_vote(out, n_burn = 1000, int = 10)
}
\arguments{
\item{out}{output from \code{neodowns}}
}
\value{
A list, which has the following components:
\item{d_voters}{a matrix containing the simulated ideological positions of voters.}
\item{d_cands}{a matrix containing the simulated ideological positions of candidates.}
}
\description{
\code{extract()} extract various quantities of interest from the output of \code{neodowns}.
}
