% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neodowns.R
\name{neodowns}
\alias{neodowns}
\title{Perform computer simulations of candidate behavior}
\usage{
neodowns(
  data,
  strategy = "max1",
  n_iter = 100,
  mu_c = 1,
  mu_b = 2,
  sigma_c = 0.5,
  sigma_b = 0.5,
  eps_sd = 0.5,
  unit = 0.05,
  seed = 14231
)
}
\arguments{
\item{data}{simulated data from \code{sim_data}}

\item{n_iter}{Number of iterations.}

\item{mu_c}{Parameter that controls for the average level of spatial voting.}

\item{mu_b}{Parameter that controls for the average level of co-ethnic voting.}

\item{sigma_c}{Parameter that controls for the standard deviation around c}

\item{sigma_b}{Parameter that controls for the standard deviation around b}

\item{eps_sd}{Parameter that controls for the level of idiosyncratic factors.}

\item{unit}{Unit with which candidates move in the two-dimensional space.}

\item{force}{Logical value for forcing moderate candidates taking more moderate positions than extreme candidates within each group.}

\item{while_max}{Number of trials to apply \code{force}}

\item{boost}{Unit by which a simulation pushes a stuck candidate out of a trap.
@return
A list, which has the following components:
\item{d_voters}{a matrix containing the simulated ideological positions of voters.}
\item{d_cands}{a matrix containing the simulated ideological positions of candidates.}}
}
\description{
\code{neodowns()} simulates the changes in candidate positions under three electoral strategies.
}
