% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_burnin.R
\name{plot_burnin}
\alias{plot_burnin}
\title{Plot candiate behavior inside and outside the burnin period}
\usage{
plot_burnin(out)
}
\arguments{
\item{out}{Output from \code{neodowns}
@return}
}
\description{
\code{plot_burnin()} visualizes the sequence of candiate positions.
}
