% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot candidate behavior inside and outside the burn-in period}
\usage{
plot_density(out)
}
\arguments{
\item{out}{Output from \code{neodowns}}
}
\value{
ggplot output
}
\description{
\code{plot_trace()} makes the trace plot for one or more chains.
}
