% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R, R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot candidate behavior inside and outside the burn-in period}
\usage{
plot_trace(out)

plot_trace(out)
}
\arguments{
\item{out}{Output from \code{neodowns}}
}
\value{
ggplot output

ggplot output
}
\description{
\code{plot_trace()} makes the trace plot for one or more chains.

\code{plot_trace()} makes the trace plot for one or more chains.
}
