% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Simulate voter and candidate data}
\usage{
see(data)
}
\arguments{
\item{data}{Simulated data.}
}
\description{
\code{sim_data()} simulates ideological positions of voters and candidates.
}
