% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate voter and candidate data}
\usage{
sim_data(
  N_voters = 1000,
  N_groups = 3,
  skew = 2,
  dev = 0.75,
  dist = "Normal",
  init = 1,
  eth = 1,
  n_candidates = 3
)
}
\arguments{
\item{N_voters}{Number of voters in the simulation.}

\item{N_groups}{Number of groups in the simulation.}

\item{skew}{Skew---s parameter---of the zipf distribution for generating relative balance of ethnic groups,
a high skew implies the biggest group is much bigger then other groups while
a low skew implies more balance,
.6 ensures no group has an outright majority.}

\item{dev}{Parameter that tweaks the variance of the various bits of code that generate voter ideological positions,
a high dev implies more variance within groups, a low dev implies less variance within groups.}

\item{dist}{Voter distribution, supporting 'Normal,' 'Flat,' 'Polarized,' or 'Clustered'}

\item{init}{'initial distance' between candidates/groups.
a high init parameter indicates should form a circle around the origin with each group far away from it, whereas
a low init parameter implies groups clustered around the origin with significant overlap.}

\item{eth}{Parameter that controls for the separation of ideological space by ethnicity.
eth = 0 is the maximum and eth = 1 is the minimum separation.}

\item{seed}{Random seed for reproducibility.}
}
\description{
\code{sim_data()} simulates ideological positions of voters and candidates.
}
