% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_burnin.R
\name{plot_burnin}
\alias{plot_burnin}
\title{Plot candidate behavior inside and outside the burn-in period}
\usage{
plot_burnin(out, n_iter, n_show = 20)
}
\arguments{
\item{out}{Output from \code{neodowns}}

\item{n_iter}{Number of iterations to plot}

\item{n_show}{Number of points to plot}
}
\value{
ggplot output
}
\description{
\code{plot_burnin()} visualizes the sequence of candidate positions.
}
